move test and print_version to utils.py
move load_notebook to utils.py

move Session to utils.py

clean up __init__.py
